refactor(plus): extract Litestream into a shared bencher_litestream crate
Move the Litestream config types, YAML rendering, restore/replicate
runtime, and the autocheckpoint PRAGMA out of bencher_json and the API
server into a new Bencher Plus crate, plus/bencher_litestream, so the
code can be shared by path with other consumers.

- bencher_json re-exports the config types from the new crate at their
  existing paths and keeps the From<LogLevel> conversion, so the public
  API and the OpenAPI schema are unchanged.
- services/api calls bencher_litestream::run_litestream, computing the
  database path, config path, and log level at the call site.
- bencher_config uses bencher_litestream::DISABLE_AUTOCHECKPOINT_PRAGMA.

The crate is feature-gated so config-only consumers stay light: the
bencher_valid backend is selected via server/client features (threaded
from bencher_json, not hardcoded), and yaml/runtime/schema are opt-in.
The moved unit tests pass and cargo gen-spec / gen-ts produce no diff.
