+++ services/api/Dockerfile
@@ -68,6 +68,7 @@ COPY plus/bencher_github_client bencher_github_client
 COPY plus/bencher_google_client bencher_google_client
 COPY plus/bencher_google_index bencher_google_index
 COPY plus/bencher_license bencher_license
+COPY plus/bencher_litestream bencher_litestream
 COPY plus/bencher_oci_storage bencher_oci_storage
 COPY plus/bencher_otel bencher_otel
 COPY plus/bencher_otel_provider bencher_otel_provider
